Moved Billing models to Billing namespace
