+++ src/Api/Billing/Models/Responses/BillingResponseModel.cs
@@ -1,8 +1,8 @@
-﻿using Bit.Core.Enums;
+﻿using Bit.Core.Billing.Models;
+using Bit.Core.Enums;
 using Bit.Core.Models.Api;
-using Bit.Core.Models.Business;
 
-namespace Bit.Api.Models.Response;
+namespace Bit.Api.Billing.Models.Responses;
 
 public class BillingResponseModel : ResponseModel
 {
